[IMP] refactor test
