+++ sale_coupon_advanced/models/sale_order.py
@@ -7,6 +7,12 @@ class SaleOrder(models.Model):
     _inherit = "sale.order"
 
     def _get_applicable_no_code_promo_program(self):
+        programs = super()._get_applicable_no_code_promo_program()
+        filtered_programs = self._filter_cumulative_programs(programs)
+        filtered_programs = self._filter_pricelist_programs(filtered_programs)
+        return self.env["sale.coupon.program"].browse(filtered_programs)
+
+    def _filter_cumulative_programs(self, programs):
         # filter recordset of programs to cut non cumulative programs
         # method should respect the initial sequence of programs
         programs = super()._get_applicable_no_code_promo_program()
